move hook preprocess_field to TripalEntityHooks
